+++ Aprendizado/Controllers/CidadeController.cs
@@ -2,12 +2,14 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
+using System.Web.Security;
 using System.Web.Mvc;
 using Aprendizado.Entity;
 using Aprendizado.Models;
 
 namespace Aprendizado.Controllers
 {
+    [Authorize]
     public class CidadeController : Controller
     {
         private CidadeModel cidadeModel = new CidadeModel();
@@ -20,57 +22,65 @@ namespace Aprendizado.Controllers
 
         public ActionResult Edit(int id)
         {
-            Cidade t = new Cidade();
-            ViewBag.Titulo = "Nova Cidade";
-
-            string idEstadoSelecionado = "MG";
-
-            if (id != 0)
+            if (Roles.IsUserInRole(User.Identity.Name, "Administrador"))
             {
-                t = cidadeModel.obterCidade(id);
-                idEstadoSelecionado = t.UF;
-                ViewBag.Titulo = "Editar Cidade";
+                Cidade t = new Cidade();
+                ViewBag.Titulo = "Nova Cidade";
+
+                string idEstadoSelecionado = "MG";
+
+                if (id != 0)
+                {
+                    t = cidadeModel.obterCidade(id);
+                    idEstadoSelecionado = t.UF;
+                    ViewBag.Titulo = "Editar Cidade";
+                }
+
+                ViewBag.UF
+                    = new SelectList(estadoModel.todosEstados(),
+                        "UF", "Descricao", idEstadoSelecionado);
+
+
+                return View(t);
             }
-
-            ViewBag.UF
-                = new SelectList(estadoModel.todosEstados(),
-                    "UF", "Descricao", idEstadoSelecionado);
-
-
-            return View(t);
+            return Redirect("/Shared/Restrito");
         }
 
         [HttpPost]
         public ActionResult Edit(Cidade t, Estado e)
         {
-            ViewBag.UF
-                = new SelectList(estadoModel.todosEstados(),
-                    "UF", "Descricao", e);
+            if (Roles.IsUserInRole(User.Identity.Name, "Administrador"))
+            {
+                ViewBag.UF
+                    = new SelectList(estadoModel.todosEstados(),
+                        "UF", "Descricao", e);
 
-            if (!validarCidade(t))
-            {
-                ViewBag.Erro = "Erro na validação do Cidade";
-                return View(t);
-            }
+                if (!validarCidade(t))
+                {
+                    ViewBag.Erro = "Erro na validação do Cidade";
+                    return View(t);
+                }
 
-            string erro = null;
-            if (t.idCidade == 0)
-            {
-                erro = cidadeModel.adicionarCidade(t);
+                string erro = null;
+                if (t.idCidade == 0)
+                {
+                    erro = cidadeModel.adicionarCidade(t);
+                }
+                else
+                {
+                    erro = cidadeModel.editarCidade(t);
+                }
+                if (erro == null)
+                {
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    ViewBag.Erro = erro;
+                    return View(t);
+                }
             }
-            else
-            {
-                erro = cidadeModel.editarCidade(t);
-            }
-            if (erro == null)
-            {
-                return RedirectToAction("Index");
-            }
-            else
-            {
-                ViewBag.Erro = erro;
-                return View(t);
-            }
+            return Redirect("/Shared/Restrito");
         }
 
         private bool validarCidade(Cidade c)
@@ -85,9 +95,15 @@ namespace Aprendizado.Controllers
 
         public ActionResult Delete(int id)
         {
-            Cidade t = cidadeModel.obterCidade(id);
-            cidadeModel.excluirCidade(t);
-            return RedirectToAction("Index");
+            if (Roles.IsUserInRole(User.Identity.Name, "Administrador"))
+            {
+                Cidade t = cidadeModel.obterCidade(id);
+                cidadeModel.excluirCidade(t);
+                return RedirectToAction("Index");
+            }
+            return Redirect("/Shared/Restrito");
+
+
         }
 
     }
