+++ Aprendizado/Controllers/TipoEnderecoController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
+using System.Web.Security;
 using System.Web.Mvc;
 using Aprendizado.Entity;
 using Aprendizado.Models;
@@ -26,42 +27,51 @@ namespace Aprendizado.Controllers
 
         public PartialViewResult Edit(int id)
         {
-            TipoEndereco te = new TipoEndereco();
-            if (id != 0)
-            {
-                te = teModel.obterTipoEndereco(id);
-            }
-            return PartialView(te);
+            
+                TipoEndereco te = new TipoEndereco();
+                if (id != 0)
+                {
+                    te = teModel.obterTipoEndereco(id);
+                }
+                return PartialView(te);
         }
 
         [HttpPost]
         public ActionResult Edit(TipoEndereco te)
         {
-            string erro = null;
-            if (te.idTipoEndereco == 0)
+            if (Roles.IsUserInRole(User.Identity.Name, "Administrador"))
             {
-                erro = teModel.adicionarTipoEndereco(te);
+                string erro = null;
+                if (te.idTipoEndereco == 0)
+                {
+                    erro = teModel.adicionarTipoEndereco(te);
+                }
+                else
+                {
+                    erro = teModel.editarTipoEndereco(te);
+                }
+                if (erro == null)
+                {
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    ViewBag.Erro = erro;
+                    return View(te);
+                }
             }
-            else
-            {
-                erro = teModel.editarTipoEndereco(te);
-            }
-            if (erro == null)
-            {
-                return RedirectToAction("Index");
-            }
-            else
-            {
-                ViewBag.Erro = erro;
-                return View(te);
-            }
+            return Redirect("/Shared/Restrito");
         }
 
         public ActionResult Delete(int id)
         {
-            TipoEndereco te = teModel.obterTipoEndereco(id);
-            teModel.excluirTipoEndereco(te);
-            return RedirectToAction("Index");
+            if (Roles.IsUserInRole(User.Identity.Name, "Administrador"))
+            {
+                TipoEndereco te = teModel.obterTipoEndereco(id);
+                teModel.excluirTipoEndereco(te);
+                return RedirectToAction("Index");
+            }
+            return Redirect("/Shared/Restrito");
         }
 
     }
